[BoundsSafety] Reland the miscompilation fix for lvalues binding with temporary locations (#11469)
- Solution from #10969:

This should fix a miscompilation bug found during adoption. When a base is lvalue and is wrapped in OVE, the compiler created a temporary and incorrectly mapped the temporary's address as the lvalue instead of the actual address of the lvalue.

The solution is to emit lvalue directly instead of creating extra temporary.

- Issue with the solution:

The issue #10969 got originally reverted due "error: cannot compile this l-value expression yet" occurred when __terminated_by_to_indexable is wrapped in an OVE.

- Fix for the new issue with __terminated_by_to_indexable:

The fix is to implement "TerminatedByToIndexableExprLValue" so the codegen can also emit as if it's an lvalue.

rdar://146329029
(cherry picked from commit 04da1af71ca7987620909b31fcdfc7934c1f04a7)