[Sema] Fix parameter index checks on explicit object member functions (#165586)
With the C++23 explicit object parameter feature, it is no longer
sufficient to only check if a function is an instance method to
determine if it has an implicit this argument. That causes problems in
attributes that have parameter indexes.

(cherry picked from commit 1fe3b2d81bd3061f1989e9677394f25725132301)